@@ -1,20 +1,18 @@
-import {NgModule} from '@angular/core';
+import {APP_INITIALIZER, NgModule} from '@angular/core';
 import {BrowserModule} from '@angular/platform-browser';
 import {AppRoutingModule} from './app-routing.module';
 import {AppComponent} from './app.component';
 import {BrowserAnimationsModule} from '@angular/platform-browser/animations';
-import {LoginComponent} from './components/login/login.component';
 import {MatSidenavModule} from "@angular/material/sidenav";
 import {MatIconModule} from "@angular/material/icon";
 import {MatToolbarModule} from "@angular/material/toolbar";
 import {AboutComponent} from './components/about/about.component';
 import {CaseExplorerComponent} from './components/case-explorer/case-explorer.component';
-import {HttpClientModule} from "@angular/common/http";
+import {HTTP_INTERCEPTORS, HttpClientModule} from "@angular/common/http";
 import {MatSortModule} from "@angular/material/sort";
 import {RegistryViewerComponent} from './components/registry-viewer/registry-viewer.component';
 import {FormsModule, ReactiveFormsModule} from "@angular/forms";
-import {MatDatepickerModule} from "@angular/material/datepicker";
-import {MatNativeDateModule} from "@angular/material/core";
+import {DateAdapter, MAT_DATE_FORMATS, MatNativeDateModule} from "@angular/material/core";
 import { DemographicDataComponent } from './components/registry-viewer/demographic-data/demographic-data.component';
 import { SummaryViewComponent } from './components/registry-viewer/summary-view/summary-view.component';
 import { ChronologicalViewComponent } from './components/registry-viewer/chronological-view/chronological-view.component';
@@ -26,7 +24,7 @@ import { AnnotationsComponent } from './components/registry-viewer/side-panel/an
 import {DrawerService} from "./service/drawer.service";
 import {MatMultiSortModule} from "ngx-mat-multi-sort";
 import {AddRecordDialogComponent} from './components/registry-viewer/add-record-dialog/add-record-dialog.component';
-import {DatePipe} from "@angular/common";
+import {DatePipe, NgOptimizedImage} from "@angular/common";
 import { SectionComponent } from './components/registry-viewer/summary-view/section/section.component';
 import {ScrollingModule} from "@angular/cdk/scrolling";
 import { ConformationDialogComponent } from './components/conformation-dialog/conformation-dialog.component';
@@ -43,12 +41,26 @@ import {MatButtonToggleModule} from "@angular/material/button-toggle";
 import {MatButtonModule} from "@angular/material/button";
 import {MatSlideToggleModule} from "@angular/material/slide-toggle";
 import {DemoModeComponent} from "./components/demo-mode/demo-mode.component";
+import {ConfigService} from "./service/config.service";
+import {MatSnackBarModule} from "@angular/material/snack-bar";
+import {MatDatepickerModule} from "@angular/material/datepicker";
+import {MatTooltipModule} from "@angular/material/tooltip";
+import {APP_DATE_FORMATS, AppDateAdapter} from "./provider/format-datepicker";
+import {AuthHttpInterceptor, AuthModule} from '@auth0/auth0-angular';
+import { Auth0LoginComponent } from './auth0-login/auth0-login.component';
+import {MatMenuModule} from "@angular/material/menu";
+import {MatDividerModule} from "@angular/material/divider";
+import { LandingComponent } from './components/landing/landing.component';
+import {MatRadioModule} from "@angular/material/radio";
+
+export const configFactory = (configService: ConfigService) => {
+  return () => configService.loadConfig();
+};
 
 
 @NgModule({
   declarations: [
     AppComponent,
-    LoginComponent,
     AboutComponent,
     CaseExplorerComponent,
     RegistryViewerComponent,
@@ -64,6 +76,8 @@ import {DemoModeComponent} from "./components/demo-mode/demo-mode.component";
     SectionComponent,
     ConformationDialogComponent,
     DemoModeComponent,
+    Auth0LoginComponent,
+    LandingComponent,
   ],
   imports: [
     BrowserModule,
@@ -75,7 +89,6 @@ import {DemoModeComponent} from "./components/demo-mode/demo-mode.component";
     MatSortModule,
     HttpClientModule,
     ReactiveFormsModule,
-    MatDatepickerModule,
     MatNativeDateModule,
     MatExpansionModule,
     FormsModule,
@@ -92,9 +105,50 @@ import {DemoModeComponent} from "./components/demo-mode/demo-mode.component";
     MatProgressSpinnerModule,
     MatButtonToggleModule,
     MatButtonModule,
-    MatSlideToggleModule
+    MatSlideToggleModule,
+    MatSnackBarModule,
+    MatDatepickerModule,
+    NgOptimizedImage,
+    MatTooltipModule,
+    AuthModule.forRoot({
+        domain: 'grady-temp.us.auth0.com',
+        clientId: '4T568aTy0dj7keOCla7FubFQO7hJ9iiH',
+        authorizationParams: {
+          redirect_uri: window.location.href,
+          audience: 'http://smartchartsuite.grady/registry-viewer-api/',
+          scope: 'profile email openid read:scd read:syphilis write:metadata write:scd write:syphilis'
+        },
+        httpInterceptor: {
+          allowedList: [{
+            uri: 'https://smartchartsuite.dev.heat.icl.gtri.org/registry-viewer-api/*',
+            // tokenOptions: {
+            //   authorizationParams: {
+            //     audience: 'http://smartchartsuite.grady/registry-viewer-api/'
+            //   }
+            // }
+          }
+
+          ]
+        }
+      }
+    ),
+    MatMenuModule,
+    MatDividerModule,
+    MatRadioModule
   ],
-  providers: [DrawerService, DatePipe],
+  providers: [
+    DrawerService,
+    DatePipe,
+    {provide: DateAdapter, useClass: AppDateAdapter},
+    {provide: MAT_DATE_FORMATS, useValue: APP_DATE_FORMATS},
+    {
+      provide: APP_INITIALIZER,
+      useFactory: configFactory,
+      deps: [ConfigService],
+      multi: true
+    },
+    { provide: HTTP_INTERCEPTORS, useClass: AuthHttpInterceptor, multi: true },
+  ],
   bootstrap: [AppComponent]
 })
 export class AppModule {
