@@ -1,27 +1,32 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import {AboutComponent} from "./components/about/about.component";
-import {LoginComponent} from "./components/login/login.component";
 import {CaseExplorerComponent} from "./components/case-explorer/case-explorer.component";
 import {RegistryViewerComponent} from "./components/registry-viewer/registry-viewer.component";
+import { AuthGuard } from '@auth0/auth0-angular';
+import {LandingComponent} from "./components/landing/landing.component";
 
 const routes: Routes = [
   //We need to add Login guard to this section
   {
     path: '',
-    component: CaseExplorerComponent
+    component: LandingComponent,
+    canActivate: [AuthGuard],
+  },
+  {
+    path: 'case',
+    component: CaseExplorerComponent,
+    canActivate: [AuthGuard],
   },
   {
     path: 'about',
-    component: AboutComponent
+    component: AboutComponent,
+    canActivate: [AuthGuard],
   },
   {
-    path: 'login',
-    component: LoginComponent
-  },
-  {
-    path: 'registry-viewer/:id',
-    component: RegistryViewerComponent
+    path: 'case/:id',
+    component: RegistryViewerComponent,
+    canActivate: [AuthGuard],
   },
   { // This path MUST ALWAYS be the last path!!!
     // Do not add any paths below this point or they will not work and will be redirected to landing.
