@@ -6,7 +6,7 @@ import {MAT_DIALOG_DATA, MatDialogRef} from "@angular/material/dialog";
 @Component({
   selector: 'app-conformation-dialog',
   templateUrl: './conformation-dialog.component.html',
-  styleUrls: ['./conformation-dialog.component.css']
+  styleUrls: ['./conformation-dialog.component.scss']
 })
 export class ConformationDialogComponent {
   message: string = "Are you sure?"
