@@ -7,7 +7,7 @@ import {DemoModeService} from "../../../service/demo-mode.service";
 @Component({
   selector: 'app-demographic-data',
   templateUrl: './demographic-data.component.html',
-  styleUrls: ['./demographic-data.component.css']
+  styleUrls: ['./demographic-data.component.scss']
 })
 export class DemographicDataComponent implements OnInit {
 
@@ -25,7 +25,7 @@ export class DemographicDataComponent implements OnInit {
 
   ngOnInit(): void {
     const caseId =  this.route.snapshot.params['id'];
-    this.caseRecordService.searchCases([caseId.toString()], ['caseId']).subscribe(
+    this.caseRecordService.searchCases('syphilis', [caseId.toString()], ['caseId']).subscribe(
       {
       next: value => this.demographicsData = value?.data[0]
       }
