@@ -8,7 +8,7 @@ import {CaseRecordsService} from "../../../../service/case-records.service";
 @Component({
   selector: 'app-details',
   templateUrl: './details.component.html',
-  styleUrls: ['./details.component.css']
+  styleUrls: ['./details.component.scss']
 })
 export class DetailsComponent implements OnInit {
 
