@@ -1,13 +1,14 @@
-import {Component, ViewChild} from '@angular/core';
+import {Component, Input, ViewChild} from '@angular/core';
 import {MatAccordion} from "@angular/material/expansion";
 import {DrawerService} from "../../../service/drawer.service";
 
 @Component({
   selector: 'app-side-panel',
   templateUrl: './side-panel.component.html',
-  styleUrls: ['./side-panel.component.css']
+  styleUrls: ['./side-panel.component.scss']
 })
 export class SidePanelComponent {
+  @Input() registrySchemaTag: string;
 
   @ViewChild(MatAccordion) accordion: MatAccordion;
   allPanelsExpanded = true;
@@ -19,5 +20,6 @@ export class SidePanelComponent {
   onClosePanel() {
     this.sidenavService.close();
   }
+
 }
 
