@@ -2,5 +2,5 @@ import {CaseRecord} from "./case.record";
 
 export class CaseRecordApiResponse {
   count: number;
-  data: CaseRecord[]
+  data: CaseRecord[];
 }
