@@ -1,13 +1,13 @@
 import { Injectable } from '@angular/core';
 import {HttpClient, HttpParams} from "@angular/common/http";
 import {BehaviorSubject, map, Observable} from "rxjs";
-import {CaseRecordApiResponse} from "../model/case.record.api.response";
-import {CaseRecord} from "../model/case.record";
-import {environment} from "../../environments/environment";
-import {ChronologicalCaseRecord} from "../model/chronological.case.record";
-import {Annotation} from "../model/annotation";
-import {Question} from "../model/question";
+import {CaseRecordApiResponse} from "../domain/case.record.api.response";
+import {CaseRecord} from "../domain/case.record";
+import {ChronologicalCaseRecord} from "../domain/chronological.case.record";
+import {Annotation} from "../domain/annotation";
+import {Question} from "../domain/question";
 import {DemoModeService} from "./demo-mode.service";
+import {EnvironmentHandlerService} from "./environment-handler.service";
 
 @Injectable({
   providedIn: 'root'
@@ -24,16 +24,24 @@ export class CaseRecordsService {
   sections: string[] = [];
   sections$: BehaviorSubject<string[]>;
 
+  baseApiUrl: string;
+
   selectedCaseRecord: any;
   selectedCaseRecord$: BehaviorSubject<any>;
 
-  constructor(private http: HttpClient, private demoModeService: DemoModeService) {
+  constructor(private http: HttpClient, private demoModeService: DemoModeService, private environmentHandler: EnvironmentHandlerService) {
     this.caseRecordChronologicalData$ = new BehaviorSubject(this.caseRecordChronologicalData);
     this.caseRecordChronologicalDataStored$ = new BehaviorSubject(this.caseRecordChronologicalDataStored);
     this.sections$ = new BehaviorSubject(this.sections);
     this.selectedCaseRecord$ = new BehaviorSubject(this.sections);
+    this.baseApiUrl = this.environmentHandler.getBaseApiURL();
+
     this.demoModeService.latestDate$.subscribe({
-      next: latestDate => {if(latestDate){this.filterLatestDateData(latestDate, this.caseRecordChronologicalDataStored$.value)}}
+      next: latestDate => {
+        if(latestDate) {
+          this.filterLatestDateData(latestDate, this.caseRecordChronologicalDataStored$.value)
+        }
+      }
     });
   }
 
@@ -41,7 +49,7 @@ export class CaseRecordsService {
     this.selectedCaseRecord$.next(selectedCaseRecord);
   }
 
-  updateCaseRecord(keyValue: any, caseId: number, contentId?: number) : Observable<any>{
+  updateCaseRecord(registrySchemaTag: string, keyValue: any, caseId: number, contentId?: number) : Observable<any>{
 
     let params = new HttpParams().set("caseId", caseId);
 
@@ -49,15 +57,15 @@ export class CaseRecordsService {
       params = new HttpParams().set("caseId", caseId).set("contentId", contentId);
     }
 
-    return this.http.put(environment.apiUrl + 'case-record', keyValue, {params}).pipe(
+    return this.http.put(this.baseApiUrl + 'case-record/' + registrySchemaTag, keyValue, {params}).pipe(
       map((result: any) => {
-        this.getByCaseId(caseId).subscribe();
+        this.getByCaseId(registrySchemaTag, caseId).subscribe();
         }
       ),
     );
   }
 
-  searchCases(searchTerms?: string[], fieldsList? : string[]):  Observable<CaseRecordApiResponse> {
+  searchCases(registrySchemaTag: string, searchTerms?: string[], fieldsList? : string[]):  Observable<CaseRecordApiResponse> {
     let options = {};
     if(searchTerms?.length>0 ||fieldsList?.length> 0){
       const terms: string = searchTerms.join(', ');
@@ -68,11 +76,10 @@ export class CaseRecordsService {
       options = { params: httpParams };
     }
 
-    return this.http.get(environment.apiUrl + 'search-cases', options).pipe(
+    return this.http.get(this.baseApiUrl + 'search-cases/' + registrySchemaTag, options).pipe(
       map((result: any) => {
         let caseList: CaseRecord[] = result.cases.map(
           (element: any) => {
-            //TODO this parser will not be needed once the API returns correct data
             let parsedCase: CaseRecord = {
               caseId: element.caseId,
               givenName: element.firstName,
@@ -99,13 +106,13 @@ export class CaseRecordsService {
     );
   };
 
-  getByCaseId (caseId):  Observable<any> {
+  getByCaseId (registrySchema, caseId):  Observable<any> {
     let options = {};
     if(!!caseId){
       const httpParams = new HttpParams().set('caseId', caseId);
       options = { params: httpParams };
     }
-    return this.http.get(environment.apiUrl + 'case-record',  options).pipe(
+    return this.http.get(this.baseApiUrl + 'case-record/' + registrySchema,  options).pipe(
       map((result: any) => {
         const mappedCaseRecords = this.createCaseRecordChronologicalData(result);
         if(this.selectedCaseRecord$?.value?.contentId && mappedCaseRecords.length > 0){
@@ -124,12 +131,12 @@ export class CaseRecordsService {
     );
   };
 
-  getQuestions (section: string): Observable<Question[]> {
+  getQuestions (section: string, registrySchemaTag: string): Observable<Question[]> {
 
     const httpParams = new HttpParams().set('section', section);
     const options = { params: httpParams };
 
-    return this.http.get(environment.apiUrl + 'questions',  options).pipe(
+    return this.http.get(this.baseApiUrl + 'questions/' + registrySchemaTag,  options).pipe(
       map((result: any) => {
           return result;
         }
